Hw04 auth

+++ service/index.js
@@ -1,34 +1,55 @@
-const Contact = require("./schemas/contact");
+const Contact = require("../schemas/contact.schema");
+const User = require("../schemas/users.schema");
 
-const listContacts = async () => {
-  return Contact.find();
+const listContacts = async (owner) => {
+  return Contact.find({ owner });
 };
 
-const getContactById = async (contactId) => {
-  return Contact.findOne({ _id: contactId });
+const getContactById = async (contactId, owner) => {
+  return Contact.findOne({ _id: contactId, owner });
 };
 
-const removeContact = (contactId) => {
-  return Contact.findOneAndDelete({ _id: contactId });
+const removeContact = (contactId, owner) => {
+  return Contact.findOneAndDelete({ _id: contactId, owner });
 };
 
-const createContact = async ({ name, email, phone }) => {
-  return Contact.create({ name, email, phone });
+const createContact = async ({ name, email, phone }, owner) => {
+  return Contact.create({ name, email, phone, owner });
 };
 
-const updateContact = async (contactId, body) => {
-  await Contact.findByIdAndUpdate({ _id: contactId }, body);
+const updateContact = async (contactId, body, owner) => {
+  await Contact.findByIdAndUpdate({ _id: contactId, owner }, body);
 };
 
-const updateStatusContact = async (contactId, favorite) => {
-  await Contact.updateOne({ _id: contactId }, { favorite: favorite });
+const updateStatusContact = async (contactId, favorite, owner) => {
+  await Contact.updateOne({ _id: contactId, owner }, { favorite: favorite });
+};
+
+const getFavoriteContacts = (owner, favorite) => {
+  return Contact.find({ owner, favorite });
+};
+
+const getUser = (email) => {
+  return User.findOne({ email });
+};
+
+const createtUser = (email, password) => {
+  return User.create({ email, password });
+};
+
+const updateSubscription = (subscription, owner) => {
+  return User.findByIdAndUpdate({ _id: owner }, { subscription: subscription });
 };
 
 module.exports = {
   createContact,
+  createtUser,
   getContactById,
+  getFavoriteContacts,
+  getUser,
   listContacts,
   removeContact,
   updateContact,
   updateStatusContact,
+  updateSubscription,
 };
